add hasStore method to idb helper

+++ src/services/storage.service.ts
@@ -41,6 +41,10 @@ export class IDB<S extends DBSchema> {
 		});
 	}
 
+	async hasStore(name: StoreNames<S>, transaction?: IDBPReadableTransaction<S>) {
+		const db = transaction?.db ?? (await this.instance);
+		return db.objectStoreNames.contains(name);
+	}
 	async createStore(name: StoreNames<S>, transaction?: IDBPWritableTransaction<S>) {
 		const db = transaction?.db ?? (await this.instance);
 		if (db.objectStoreNames.contains(name)) return;
@@ -57,32 +61,39 @@ export class IDB<S extends DBSchema> {
 	}
 
 	async keys(store: StoreNames<S>, transaction?: IDBPReadableTransaction<S>) {
+		if (!(await this.hasStore(store, transaction))) return;
 		const tx = transaction?.objectStore(store) ?? (await this.instance).transaction(store, "readonly").objectStore(store);
 		return tx.getAllKeys();
 	}
 	async has(store: StoreNames<S>, key: StoreKey<S, StoreNames<S>>, transaction?: IDBPReadableTransaction<S>) {
+		if (!(await this.hasStore(store, transaction))) return false;
 		const tx = transaction?.objectStore(store) ?? (await this.instance).transaction(store, "readonly").objectStore(store);
 		const exists = await tx.getKey(key);
 		return !!exists;
 	}
 	async get(store: StoreNames<S>, key: StoreKey<S, StoreNames<S>>, transaction?: IDBPReadableTransaction<S>) {
+		if (!(await this.hasStore(store, transaction))) return;
 		const tx = transaction?.objectStore(store) ?? (await this.instance).transaction(store, "readonly").objectStore(store);
 		return tx.get(key);
 	}
 	async set(store: StoreNames<S>, key: StoreKey<S, StoreNames<S>>, value: StoreValue<S, StoreNames<S>>, transaction?: IDBPWritableTransaction<S>) {
+		if (!(await this.hasStore(store, transaction))) return;
 		const tx = transaction?.objectStore(store) ?? (await this.instance).transaction(store, "readwrite").objectStore(store);
 		return tx.put(value, key);
 	}
 	async delete(store: StoreNames<S>, key: StoreKey<S, StoreNames<S>>, transaction?: IDBPWritableTransaction<S>) {
+		if (!(await this.hasStore(store, transaction))) return;
 		const tx = transaction?.objectStore(store) ?? (await this.instance).transaction(store, "readwrite").objectStore(store);
 		return tx.delete(key);
 	}
 	async clear(store: StoreNames<S>, transaction?: IDBPWritableTransaction<S>) {
+		if (!(await this.hasStore(store, transaction))) return;
 		const tx = transaction?.objectStore(store) ?? (await this.instance).transaction(store, "readwrite").objectStore(store);
 		return tx.clear();
 	}
 
 	async update<T extends StoreValue<S, StoreNames<S>>, R extends StoreValue<S, StoreNames<S>> = StoreValue<S, StoreNames<S>>>(store: StoreNames<S>, key: StoreKey<S, StoreNames<S>>, updater: (value: T) => R, transaction?: IDBPWritableTransaction<S>) {
+		if (!(await this.hasStore(store, transaction))) return;
 		const tx = transaction?.objectStore(store) ?? (await this.instance).transaction(store, "readwrite").objectStore(store);
 		const current = await tx.get(key);
 		if (!current) return undefined;
